feature-extractor/requirements.txt
Flask==2.2.5
numpy==1.23.5
opencv_python==4.9.0.80
Pillow==10.4.0
torch==2.2.2
torchvision==0.17.2
